requirements/requirements.txt
# Project
minio==4.0.18
tweepy==3.7.0
yweather==0.1.1
numpy==1.16.4
pandas==0.24.2
# tensorflow==1.14.0
# tensorflow==2.0.0-beta1
# tensorflow-hub==0.5.0
# bert-tensorflow==1.0.1

django==2.2.4
dj-database-url==0.5.0
djangorestframework==3.9.1
djangorestframework-oauth
# django-heroku==0.3.1

gunicorn==19.9.0

Pillow==5.3.0
pytz==2018.7
Unipath==1.1
whitenoise==4.1.2
